refactor: 파일 및 폴더 구조 리팩터링
- auth.service.ts를 service 폴더로 이동
- InputField.tsx를 common 폴더로 이동하여 재사용 가능성 개선
- utils 폴더를 순수 함수만 포함하도록 조정하고, 비즈니스 로직은 lib 폴더에 배치
